refactor : 로직 패키지화
